Add environment variable to set proxy bind address
This patch adds the environment variable PROXY_BIND which
can be optionally set to override the default that binds
to "*".

Binding on all addresses may inadvertently expose the proxy
service when run on a node that also has public interfaces
such as a bare metal openstack infrastructure node, or a
network node.